\name{erbio_version}
\alias{erbio_version}
\alias{erbio_report_patch_version}
\title{ERBioR Version Information}
\description{Return the ERBioR development version and frozen report-layer identifier.}
\usage{
erbio_version()
erbio_report_patch_version()
}
\value{A length-one character string.}
\examples{
erbio_version()
erbio_report_patch_version()
}
\keyword{utilities}
